split subworld loading screen into its own class
